some refactoring and metadata creation

+++ myyum/rpm/models.py
@@ -1,10 +1,9 @@
-from django.db import models
+from django.db import models, IntegrityError
 from django.core.files.storage import default_storage
 from django.contrib.auth.models import User
 
-from pyrpm import yum, rpm
-
 from myyum.rpm.fields import *
+from myyum.rpm.tools import *
 
 class Repository(models.Model):
     owner = models.ForeignKey(User)
@@ -18,7 +17,19 @@ class Repository(models.Model):
         unique_together = (
             ("owner", "name"),
         )
-
+    
+    
+    @property
+    def repodir(self):
+        return "%s/%s" % (self.owner.username, self.name)
+    
+    def update_metadata(self):
+        yum_repo = CloudYumRepository(self.repodir)
+        
+        for pkg in self.packages.all():
+            yum_repo.add_package(pkg.get_yum_package)
+        
+        yum_repo.save()
 
 class RPMPackage(models.Model):
     repository = models.ForeignKey(Repository, related_name="packages")
@@ -26,14 +37,6 @@ class RPMPackage(models.Model):
     # general information
     pkgid = models.CharField(max_length=100)
     name = models.CharField(max_length=255)
-    #summary = models.TextField(max_length=255)
-    #description = models.TextField()
-    #architecture = models.CharField(max_length=20)
-    
-    # version information
-    #version = models.CharField(max_length=20)
-    #release = models.CharField(max_length=20)
-    #epoch = models.IntegerField(max_length=20)
     
     # storage information
     filename = models.CharField(max_length=255)
@@ -57,39 +60,43 @@ class RPMPackage(models.Model):
         except:
             return ""
     
+    @classmethod
+    def create_from_rpm_file(cls, repository, package_file):
+        # parse it
+        rpm_pkg = rpm.RPM(package_file)
+        
+        # ckech for duplicates
+        if repository.packages.filter(pkgid=rpm_pkg.checksum).exists():
+            raise IntegrityError()
+        
+        # create in db
+        pkg = cls.objects.create(repository=repository, pkgid=rpm_pkg.checksum, name=rpm_pkg.header.name)
+        for header_entry in rpm_pkg.header:
+            RPMHeader.objects.create(package=pkg, tag=header_entry.tag, value=header_entry.value)
+        
+        # upload it
+        pkg.url = default_storage.save("%s/%s" % (repository.repodir, rpm_pkg.canonical_filename), package_file)
+        pkg.save()
+        
+        # update repo
+        repository.update_metadata()
+        
+        # return the new package
+        return pkg
+    
     def delete(self, *args, **kwargs):
         # check for file and delete it if existing
-        if default_storage.exists(self.url):
+        if self.url and default_storage.exists(self.url):
             default_storage.delete(self.url)
+        
+        # call super
         super(RPMPackage, self).delete(*args, **kwargs)
+        
+        # update repo
+        self.repository.update_metadata()
     
 class RPMHeader(models.Model):
     package = models.ForeignKey(RPMPackage, related_name='headers')
     
     tag = models.IntegerField()
     value = PickledObjectField()
-
-class YumPackageAdapter(yum.YumPackage):
-    """ Adapter for using the RPM header entries from
-    the database instead from a file"""
-    
-    def __init__(self, rpm_package):
-        self.binary = True
-        self.source = False
-        self.header = rpm.Header(None)
-        self.signature = rpm.Signature(None)
-        self.filelist = []
-        self.changelog = []
-        
-        self.provides = []
-        self.requires = []
-        self.obsoletes = []
-        self.conflicts = []
-        
-        self.checksum = rpm_package.pkgid
-        
-        # load header entries from db
-        for header_entry in rpm_package.headers.all():
-            self.header.entries.append(rpm.Entry(tag=header_entry.tag, value=header_entry.value))
-        
-        self._match_composite()
